+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased changes
 - Add `GetBakersRewardPeriod` to GRPCV2 API. Provided a block, then it returns information about bakers
   for the reward period of the block.
-
 - Fix a bug in how the last timeout certificate is recovered at start-up.
+- Add endpoint `GetBlockCertificates` to GRPCV2 API. Provided a block, then it returns 
+  quorum certificate, timeout certificate and epoch finalization entry contained in the block (where present).
 
 ## 6.0.4